requirements.txt
pika==1.2.0

pydantic~=2.6.1
pydantic-settings~=2.1.0
kafka-python~=2.0.2

